Revert "newtonsoft json added"
This reverts commit 03a10acbc67e6d129601b4c0abb8ab6a75e8ed1a.

+++ Models/SignUpModel.cs
@@ -9,7 +9,7 @@ namespace SplitExpenses.Models
     {
         public string Name { get; set; }
         public string UserName { get; set; }
-        public long Mobile { get; set; }
+        public int Mobile { get; set; }
         public string Email { get; set; }
         public string Address { get; set; }
         public string Password { get; set; }
